[#11] Add supplemental result to Lemma A3
I'm going to need to make repeated use of a fairly straightforward
but not trivially obvious result in Lemma A4's proof, so I'm
establishing it as part of Lemma A3. The result uses an argument
that is basically identical to other parts of Lemma A3 so it should
make sense here.